separated out base service and library type code

+++ run_tests.sh
@@ -5,6 +5,10 @@ if [ ! -d 'env' ]; then
     echo "setting up virtualenv"
     python3 -m virtualenv env
 fi
-env/bin/pip -q install -r requirements.txt
-env/bin/pip -q install -r requirements-testing.txt
-env/bin/python -m tornado.testing discover -s tokenbrowser/test
+if [ -e requirements.txt ]; then
+    env/bin/pip -q install -r requirements.txt
+fi
+if [ -e requirements-testing.txt ]; then
+    env/bin/pip -q install -r requirements-testing.txt
+fi
+env/bin/python -m tornado.testing discover -s tokenid/test
